add widening check to v2
